Extract hub logic
