+++ components/types/types.ts
@@ -1140,14 +1140,11 @@ export interface CreateFromParamsBody {
     }
 }
 
-export interface CommonSelectScreenConfig {
-    template: unknown
-    data?: {
-        Unlocked: string[]
-        Contract: MissionManifest
-        OrderedUnlocks: Unlockable[]
-        UserCentric: UserCentricContract
-    }
+export interface SelectEntranceOrPickupData {
+    Unlocked: string[]
+    Contract: MissionManifest
+    OrderedUnlocks: Unlockable[]
+    UserCentric: UserCentricContract
 }
 
 export type CompiledIoiStatemachine = unknown
